First Update
Header is responsive and i have managed to make a start to the form section of the webpage. I had some trouble with the responsiveness off the header, as i forgot about Importance in CSS. as my original declorations were below the media queries so the font size was not changing on the viewport.